don't inject for diagnostic vdocs; rename to `diagnostics.ts`
